refactor(todo): Organize components into features folder
- Moved `TodoForm` to `features/`
- Moved `TodoList` and `TodoListItem` to `features/TodoList/`
- Updated the import paths in `App.jsx` according to the new structure
